+++ Cargo.toml
@@ -9,7 +9,7 @@ members = [
 ]
 
 [workspace.package]
-version = "0.35.3"
+version = "0.35.4"
 
 [workspace.lints.clippy]
 cargo =  { level = "warn", priority = -1 }
