+++ opsqueue/src/producer/client.rs
@@ -269,11 +269,7 @@ impl InternalProducerClientError {
             Self::HTTPClientError(inner) => {
                 // Failures in which a connection could not be established are ephemeral,
                 // as these can be caused by network failures, so we want to retry:
-                if inner.is_connect()
-                    || inner.is_timeout()
-                    || inner.is_decode()
-                    || inner.is_request()
-                {
+                if inner.is_connect() || inner.is_timeout() || inner.is_decode() {
                     true
                 } else if inner.is_status() {
                     // Failures where the server returns a 5xx status code might indicate the server is (temporarily!) unhealthy,
